requirements.txt
Django==3.1.4
coverage~=5.3
python-telegram-bot>=13.1
python-dotenv~=0.15